Move community entities from locking routes
