+++ src/routes/community/community.controller.ts
@@ -2,13 +2,13 @@ import { PaginationDataDecorator } from '@/routes/common/decorators/pagination.d
 import { RouteUrlDecorator } from '@/routes/common/decorators/route.url.decorator';
 import { PaginationData } from '@/routes/common/pagination/pagination.data';
 import { CommunityService } from '@/routes/community/community.service';
-import { CampaignRank } from '@/routes/locking/entities/campaign-rank.entity';
-import { CampaignRankPage } from '@/routes/locking/entities/campaign-rank.page.entity';
-import { Campaign } from '@/routes/locking/entities/campaign.entity';
-import { CampaignPage } from '@/routes/locking/entities/campaign.page.entity';
-import { LockingEventPage } from '@/routes/locking/entities/locking-event.page.entity';
-import { LockingRank } from '@/routes/locking/entities/locking-rank.entity';
-import { LockingRankPage } from '@/routes/locking/entities/locking-rank.page.entity';
+import { CampaignRank } from '@/routes/community/entities/campaign-rank.entity';
+import { CampaignRankPage } from '@/routes/community/entities/campaign-rank.page.entity';
+import { Campaign } from '@/routes/community/entities/campaign.entity';
+import { CampaignPage } from '@/routes/community/entities/campaign.page.entity';
+import { LockingEventPage } from '@/routes/community/entities/locking-event.page.entity';
+import { LockingRank } from '@/routes/community/entities/locking-rank.entity';
+import { LockingRankPage } from '@/routes/community/entities/locking-rank.page.entity';
 import { AddressSchema } from '@/validation/entities/schemas/address.schema';
 import { ValidationPipe } from '@/validation/pipes/validation.pipe';
 import { Controller, Get, Param } from '@nestjs/common';
