+++ src/routes/community/entities/campaign-rank.entity.ts
