+++ src/routes/community/entities/campaign.page.entity.ts
@@ -1,5 +1,5 @@
 import { Page } from '@/routes/common/entities/page.entity';
-import { Campaign } from '@/routes/locking/entities/campaign.entity';
+import { Campaign } from '@/routes/community/entities/campaign.entity';
 import { ApiProperty } from '@nestjs/swagger';
 
 export class CampaignPage extends Page<Campaign> {
