+++ src/routes/community/entities/locking-rank.page.entity.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Page } from '@/routes/common/entities/page.entity';
-import { LockingRank } from '@/routes/locking/entities/locking-rank.entity';
+import { LockingRank } from '@/routes/community/entities/locking-rank.entity';
 
 export class LockingRankPage extends Page<LockingRank> {
   @ApiProperty({ type: LockingRank })
